Disable config switch (#191)
Add a config environment variable to disable the collection of metrics through this extension. Owing to the rules the lambda engine places on extensions, this isn't as simple as "just exit" - you still have to Register as an extension, and ask for at least the shutdown event. Additionally, add a config switch for "extension name" to make a complicated workaround for disabling the opentelemetry collector possible.

Co-authored-by: Pablo Collins <pablo.collins@gmail.com>